backend/requirements.txt
fastapi
uvicorn
motor
pymongo
google-genai
numpy
python-bidi
pydantic-settings
python-dotenv
